[FIX] : Failed prop type: Invalid prop  supplied to childrenForwardRef(Modal) - solved by using Bar2
